+++ src/backersManager/index.ts
@@ -1,0 +1,6 @@
+export { handleBackerRewardsOptedOut } from "./backerRewardsOptedOut";
+export { handleBackerRewardsOptedIn } from "./backerRewardsOptedIn";
+export { handleUpgraded } from "./upgraded";
+export { handleNewAllocation } from "./newAllocation";
+export { handleRewardDistributionStarted } from "./rewardDistributionStarted";
+export { handleRewardDistributionFinished } from "./rewardDistributionFinished";
