+++ src/backersManager/v1/index.ts
@@ -1,0 +1,2 @@
+export { handleGaugeCreated } from "./gaugeCreated";
+export { handleWhitelisted } from "./whitelisted";
