+++ src/builderRegistry/communityApproved.ts
@@ -1,0 +1,6 @@
+import { CommunityApproved as CommunityApprovedEvent } from "../../generated/BuilderRegistryRootstockCollective/BuilderRegistryRootstockCollective";
+import { communityApproved } from "./utils";
+
+export function handleCommunityApproved(event: CommunityApprovedEvent): void {
+  communityApproved(event.params.builder_);
+}
