+++ src/builderRegistry/index.ts
@@ -1,0 +1,12 @@
+export { handleBuilderInitialized } from "./builderInitialized";
+export { handleCommunityApproved } from "./communityApproved";
+export { handleCommunityBanned } from "./communityBanned";
+export { handleGaugeCreated } from "./gaugeCreated";
+export { handleKYCApproved } from "./kycApproved";
+export { handleKYCPaused } from "./kycPaused";
+export { handleKYCResumed } from "./kycResumed";
+export { handleKYCRevoked } from "./kycRevoked";
+export { handleRewardReceiverUpdated } from "./rewardReceiverUpdated";
+export { handleSelfPaused } from "./selfPaused";
+export { handleSelfResumed } from "./selfResumed";
+export { handleUpgraded } from "./upgraded";
