+++ src/builderRegistry/kycPaused.ts
@@ -1,0 +1,6 @@
+import { KYCPaused as KYCPausedEvent } from "../../generated/BuilderRegistryRootstockCollective/BuilderRegistryRootstockCollective";
+import { kycPaused } from "./utils";
+
+export function handleKYCPaused(event: KYCPausedEvent): void {
+  kycPaused(event.params.builder_);
+}
