+++ src/builderRegistry/rewardReceiverUpdated.ts
@@ -1,0 +1,8 @@
+import { RewardReceiverUpdated as RewardReceiverUpdatedEvent } from "../../generated/BuilderRegistryRootstockCollective/BuilderRegistryRootstockCollective";
+import { rewardReceiverUpdated } from "./utils";
+
+export function handleRewardReceiverUpdated(
+  event: RewardReceiverUpdatedEvent
+): void {
+  rewardReceiverUpdated(event.params.builder_, event.params.newRewardReceiver_);
+}
