+++ src/builderRegistry/selfPaused.ts
@@ -1,0 +1,6 @@
+import { SelfPaused as SelfPausedEvent } from "../../generated/BuilderRegistryRootstockCollective/BuilderRegistryRootstockCollective";
+import { selfPaused } from "./utils";
+
+export function handleSelfPaused(event: SelfPausedEvent): void {
+  selfPaused(event.params.builder_);
+}
