+++ src/builderRegistry/utils/index.ts
@@ -1,0 +1,13 @@
+export {
+  gaugeCreated,
+  builderInitialized,
+  communityApproved,
+  communityBanned,
+  kycApproved,
+  kycPaused,
+  kycResumed,
+  kycRevoked,
+  rewardReceiverUpdated,
+  selfPaused,
+  selfResumed,
+} from "./shared";
