+++ src/builderRegistry/v2/communityApproved.ts
@@ -1,0 +1,6 @@
+import { CommunityApproved as CommunityApprovedEvent } from "../../../generated/BuilderRegistryRootstockCollectiveV2/BuilderRegistryRootstockCollectiveV2";
+import { communityApproved } from "../utils";
+
+export function handleCommunityApproved(event: CommunityApprovedEvent): void {
+  communityApproved(event.params.builder_);
+}
