+++ src/builderRegistry/v2/dewhitelisted.ts
@@ -1,0 +1,6 @@
+import { Dewhitelisted as DewhitelistedEvent } from "../../../generated/BuilderRegistryRootstockCollectiveV2/BuilderRegistryRootstockCollectiveV2";
+import { communityBanned } from "../utils";
+
+export function handleDewhitelisted(event: DewhitelistedEvent): void {
+  communityBanned(event.params.builder_);
+}
