+++ src/builderRegistry/v2/gaugeCreated.ts
@@ -1,0 +1,6 @@
+import { GaugeCreated as GaugeCreatedEvent } from "../../../generated/BuilderRegistryRootstockCollectiveV2/BuilderRegistryRootstockCollectiveV2";
+import { gaugeCreated } from "../utils";
+
+export function handleGaugeCreated(event: GaugeCreatedEvent): void {
+  gaugeCreated(event.params.builder_, event.params.gauge_);
+}
