+++ src/builderRegistry/v2/index.ts
@@ -1,0 +1,11 @@
+export { handleBuilderActivated } from "./builderActivated";
+export { handleBuilderRewardReceiverReplacementApproved } from "./builderRewardReceiverReplacementApproved";
+export { handleCommunityApproved } from "./communityApproved";
+export { handleDewhitelisted } from "./dewhitelisted";
+export { handleGaugeCreated } from "./gaugeCreated";
+export { handleKYCApproved } from "./kycApproved";
+export { handleKYCRevoked } from "./kycRevoked";
+export { handlePaused } from "./paused";
+export { handleUnpaused } from "./unpaused";
+export { handleRevoked } from "./revoked";
+export { handlePermitted } from "./permitted";
