+++ src/builderRegistry/v2/kycApproved.ts
@@ -1,0 +1,6 @@
+import { KYCApproved as KYCApprovedEvent } from "../../../generated/BuilderRegistryRootstockCollectiveV2/BuilderRegistryRootstockCollectiveV2";
+import { kycApproved } from "../utils";
+
+export function handleKYCApproved(event: KYCApprovedEvent): void {
+  kycApproved(event.params.builder_);
+}
