+++ src/builderRegistry/v2/paused.ts
@@ -1,0 +1,6 @@
+import { Paused as PausedEvent } from "../../../generated/BuilderRegistryRootstockCollectiveV2/BuilderRegistryRootstockCollectiveV2";
+import { kycPaused } from "../utils";
+
+export function handlePaused(event: PausedEvent): void {
+  kycPaused(event.params.builder_);
+}
