+++ src/builderRegistry/v2/revoked.ts
@@ -1,0 +1,6 @@
+import { Revoked as RevokedEvent } from "../../../generated/BuilderRegistryRootstockCollectiveV2/BuilderRegistryRootstockCollectiveV2";
+import { selfPaused } from "../utils";
+
+export function handleRevoked(event: RevokedEvent): void {
+  selfPaused(event.params.builder_);
+}
