+++ src/builderRegistry/v2/unpaused.ts
@@ -1,0 +1,6 @@
+import { Unpaused as UnpausedEvent } from "../../../generated/BuilderRegistryRootstockCollectiveV2/BuilderRegistryRootstockCollectiveV2";
+import { kycResumed } from "../utils";
+
+export function handleUnpaused(event: UnpausedEvent): void {
+  kycResumed(event.params.builder_);
+}
