+++ src/gauge/backerRewardsClaimed.ts
@@ -1,16 +1,13 @@
 import {
   BackerRewardsClaimed as BackerRewardsClaimedEvent,
-  BuilderRewardsClaimed as BuilderRewardsClaimedEvent,
-} from "../generated/templates/GaugeRootstockCollective/GaugeRootstockCollective";
+} from "../../generated/templates/GaugeRootstockCollective/GaugeRootstockCollective";
 import {
   Backer,
   BackersRewardsClaimed,
-  Builder,
-  BuilderRewardsClaimed,
   BackerToBuilder,
   BackerToBuilderRewardsClaimed,
   GaugeToBuilder,
-} from "../generated/schema";
+} from "../../generated/schema";
 import { BigInt } from "@graphprotocol/graph-ts";
 
 export function handleBackerRewardsClaimed(
@@ -44,8 +41,9 @@ function _handleBackerToBuilder(event: BackerRewardsClaimedEvent): void {
   const gaugeToBuilder = GaugeToBuilder.load(event.address);
   if (gaugeToBuilder == null) return;
 
-
-  const backerToBuilderId = event.params.backer_.concat(gaugeToBuilder.builder_);
+  const backerToBuilderId = event.params.backer_.concat(
+    gaugeToBuilder.builder_
+  );
   const backerToBuilder = BackerToBuilder.load(backerToBuilderId);
   if (backerToBuilder == null) return;
 
@@ -61,26 +59,4 @@ function _handleBackerToBuilder(event: BackerRewardsClaimedEvent): void {
 
   rewardsClaimed.save();
   backerToBuilder.save();
-}
-
-export function handleBuilderRewardsClaimed(
-  event: BuilderRewardsClaimedEvent
-): void {
-  const builder = Builder.load(event.params.builder_);
-  if (builder == null) return;
-
-  const rewardsClaimedId = event.params.builder_.concat(
-    event.params.rewardToken_
-  );
-  let rewardsClaimed = BuilderRewardsClaimed.load(rewardsClaimedId);
-  if (rewardsClaimed == null) {
-    rewardsClaimed = new BuilderRewardsClaimed(rewardsClaimedId);
-    rewardsClaimed.amount_ = BigInt.zero();
-  }
-  rewardsClaimed.builder_ = event.params.builder_;
-  rewardsClaimed.token_ = event.params.rewardToken_;
-  rewardsClaimed.amount_ = rewardsClaimed.amount_.plus(event.params.amount_);
-
-  rewardsClaimed.save();
-  builder.save();
 }
