+++ src/gauge/index.ts
@@ -1,0 +1,2 @@
+export { handleBackerRewardsClaimed } from "./backerRewardsClaimed";
+export { handleBuilderRewardsClaimed } from "./builderRewardsClaimed";
