+++ src/rewardDistributor/index.ts
@@ -1,0 +1,1 @@
+export { handleUpgraded } from "./upgraded";
